dataplatform/projects/cloud-control-center/api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
authlib>=1.3.0
itsdangerous>=2.1.0
python-dotenv>=1.0.0

